jiayue-cloud-functions/temp/phl_property_data_pipeline/extract_data/requirements.txt
functions_framework==3.8.2
google-cloud-storage==3.1.0
python-dotenv==1.1.0
Requests==2.32.3
